Refactoring codice
