+++ generateID.h
@@ -1,0 +1,6 @@
+#ifndef GENERATEID_H_INCLUDED
+#define GENERATEID_H_INCLUDED
+
+long long generateUniqueID();
+
+#endif // GENERATEID_H_INCLUDED
